backend/requirements.txt
Django==4.2.7
django-allauth==0.57.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
stripe==9.12.0
requests==2.31.0
django-ratelimit==4.1.0
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
hiredis==2.3.2
django-redis==5.4.0
gunicorn==21.2.0
psutil==5.9.6
whitenoise==6.6.0
boto3==1.34.14
django-storages==1.14.2
django-debug-toolbar==4.2.0
user-agents==2.2.0
reportlab==4.0.8
django-filter==23.5